refactor(database): cleanup from /simplify review
- Export config.DefaultDatabaseTimezone and config.TimezoneDisabledSentinel
  so PG/Oracle connection layers reference the canonical constants instead
  of duplicating the magic "-" string.
- Precompute the ALTER SESSION SQL string in newTzConnector instead of
  rebuilding it (with strings.ReplaceAll + fmt.Sprintf) on every Connect.
  The connector's timezone is immutable; doing the work once at
  construction matches the per-pool-member hot-path use.
- Drop the redundant `drv` cached field on tzConnector; Driver() now
  returns inner.Driver() directly.
- Extract applyOracleKeepAlive helper to remove the duplicated keep-alive
  setup that was inlined in openTimezoneAwareOracleDB.
- Move the two TestNewConnection* dispatch tests from tz_connector_test.go
  into connection_test.go where the rest of the NewConnection coverage
  lives. tz_connector_test.go now contains only the unit tests for the
  wrapper itself.
