+++ database/oracle/tz_connector.go
@@ -14,42 +14,42 @@ import (
 // receive the same session timezone as the first connection.
 type tzConnector struct {
 	inner    driver.Connector
-	drv      driver.Driver
-	timezone string // pre-validated IANA name from config layer
+	timezone string // pre-validated IANA name; kept for error messages
+	stmt     string // precomputed ALTER SESSION SQL — never changes after construction
 }
 
 // newTzConnector returns a wrapper that applies the given timezone to every
-// new physical connection produced by inner.
+// new physical connection produced by inner. The ALTER SESSION statement is
+// built once here so Connect() (called per physical connection) doesn't
+// repeat the formatting and quote-escaping work.
 func newTzConnector(inner driver.Connector, timezone string) *tzConnector {
+	// Defensive escape — config validation rejects values containing quotes,
+	// but doubling them costs nothing and protects against future regressions
+	// in the validation layer.
+	safeTZ := strings.ReplaceAll(timezone, "'", "''")
 	return &tzConnector{
 		inner:    inner,
-		drv:      inner.Driver(),
 		timezone: timezone,
+		stmt:     fmt.Sprintf("ALTER SESSION SET TIME_ZONE = '%s'", safeTZ),
 	}
 }
 
 // Connect opens a new physical connection via the inner connector and runs
-// ALTER SESSION SET TIME_ZONE on it. If the ALTER fails the connection is
-// closed and the error bubbles up, causing database/sql to retry rather than
-// hand out a misconfigured connection.
+// the precomputed ALTER SESSION statement on it. If the ALTER fails the
+// connection is closed and the error bubbles up, causing database/sql to
+// retry rather than hand out a misconfigured connection.
 func (c *tzConnector) Connect(ctx context.Context) (driver.Conn, error) {
 	conn, err := c.inner.Connect(ctx)
 	if err != nil {
 		return nil, err
 	}
 
-	// Defensive escape — config validation already rejects values containing
-	// quotes via time.LoadLocation, but a belt-and-braces escape costs nothing
-	// and protects against future validation regressions.
-	safeTZ := strings.ReplaceAll(c.timezone, "'", "''")
-	stmt := fmt.Sprintf("ALTER SESSION SET TIME_ZONE = '%s'", safeTZ)
-
 	execer, ok := conn.(driver.ExecerContext)
 	if !ok {
 		_ = conn.Close()
 		return nil, fmt.Errorf("oracle driver.Conn does not implement driver.ExecerContext; cannot apply session timezone")
 	}
-	if _, err := execer.ExecContext(ctx, stmt, nil); err != nil {
+	if _, err := execer.ExecContext(ctx, c.stmt, nil); err != nil {
 		_ = conn.Close()
 		return nil, fmt.Errorf("apply oracle session timezone %q: %w", c.timezone, err)
 	}
@@ -57,4 +57,4 @@ func (c *tzConnector) Connect(ctx context.Context) (driver.Conn, error) {
 }
 
 // Driver returns the wrapped connector's driver, satisfying driver.Connector.
-func (c *tzConnector) Driver() driver.Driver { return c.drv }
+func (c *tzConnector) Driver() driver.Driver { return c.inner.Driver() }
